Fixed a bug where Kbar's actions don't change language when i18n does
